Cambios Alonso

+++ include/G_AccelHarmonic.h
@@ -4,6 +4,6 @@
 #include "Matrix.h"
 #include "AccelHarmonic.h"
 
-Matrix G_AccelHarmonic(Matrix* r, Matrix* U, double n_max, double m_max);
+Matrix G_AccelHarmonic(Matrix* r, Matrix* U, int n_max, int m_max);
 
 #endif
